Always render the issue form, but hidden
The issue form needs stuff in the html-head. Adding this manually only
works for core-redmine. As soon as a plugins like checklist is used, their
stuff never gets added to the html-head so they break. By always rendering
the form inline (but hidden) already, every plugin can contribute to the
html-head...

Possible optimization: when the current project is selected in the
dropdown, only show the form, no ajax required.

Refs #287